Improve browser testing
