+++ package.json
@@ -4,7 +4,7 @@
     "scripts": {
         "build": "tsc -p . && npm run copyFiles",
         "build:test:browser": "tsc -p tsconfig.browser-test.json && webpack --config webpack.config.test.js",
-        "check:tree": "ts-node ./test/utils/check-clean-tree.ts",
+        "check:tree": "ts-node ./test/commands/check-clean-tree.ts",
         "clean": "rimraf test-browser test/**/node_modules",
         "start": "node ./dist/src/main.js",
         "test": "npm-run-all unit-test integration-test rlc-integration-test validate-version-tolerance",
@@ -25,19 +25,23 @@
         "rlc-integration-test:browser": "npm run build:test:browser && karma start karma.conf.js",
         "rlc-generate-and-test": "npm-run-all -s build rlc-generate-swaggers rlc-integration-test:alone rlc-integration-test:browser",
         "rlc-integration-test:alone": "mocha -r ts-node/register --timeout 2000 './test/rlcIntegration/**/!(sampleTest).spec.ts'",
-        "start-test-server": "ts-node test/utils/start-server.ts",
+        "start-test-server": "ts-node test/commands/start-server.ts",
         "start-test-server:v2": "autorest-testserver run --appendCoverage",
         "stop-test-server": "autorest-testserver stop",
         "debug": "node --inspect-brk ./dist/src/main.js",
-        "generate-swaggers": "ts-node ./test/utils/test-swagger-gen.ts",
-        "rlc-generate-swaggers": "ts-node ./test/utils/test-swagger-gen.ts -- rlc",
+        "generate-swaggers": "ts-node ./test/commands/test-swagger-gen.ts",
+        "rlc-generate-swaggers": "ts-node ./test/commands/test-swagger-gen.ts -- rlc",
         "validate-version-tolerance": "npm run start-test-server:v2 & npm run generate-version-tolerance && npm run generate-version-tolerance:tests && npm run test-version-tolerance && npm run stop-test-server",
-        "generate-version-tolerance": "ts-node ./test/utils/version-tolerance.ts",
-        "generate-version-tolerance:tests": "ts-node ./test/utils/version-tolerance.ts --test-gen",
+        "generate-version-tolerance": "ts-node ./test/commands/version-tolerance.ts",
+        "generate-version-tolerance:tests": "ts-node ./test/commands/version-tolerance.ts --test-gen",
         "test-version-tolerance": "mocha -r ts-node/register --timeout 2000 './test/version-tolerance/**/!(sampleTest).spec.ts'",
         "clone:specs": "git clone https://github.com/Azure/azure-rest-api-specs.git ./.tmp/specs",
-        "smoke-test": "ts-node ./test/utils/smoke-test.ts",
+        "smoke-test": "ts-node ./test/commands/smoke-test.ts",
         "copyFiles": "cp src/**/*.hbs dist/ && cp src/**/**/*.hbs dist/ && cp src/generators/samples/*.hbs dist/src/generators/samples/ && cp src/generators/static/*.hbs dist/src/generators/static/ && cp src/generators/test/*.hbs dist/src/generators/test/ && cp src/restLevelClient/*.hbs dist/src/restLevelClient/"
+    },
+    "browser": {
+        "./test-browser/utils/stream-helpers.js": "./test-browser/utils/stream-helpers.browser.js",
+        "./test-browser/utils/fileSystem.js": "./test-browser/utils/fileSystem.browser.js"
     },
     "files": [
         "src",
@@ -48,7 +52,7 @@
     "dependencies": {
         "@autorest/codemodel": "^4.17.0",
         "@autorest/extension-base": "^3.4.1",
-        "@azure-rest/core-client": "1.0.0-alpha.20220112.1",
+        "@azure-rest/core-client": "1.0.0-alpha.20220415.2",
         "@azure-tools/codegen": "^2.5.294",
         "@azure/core-client": "^1.5.0",
         "@azure/core-http": "^1.2.4",
