+++ test/commands/check-clean-tree.ts
