+++ test/commands/childProcessOnExit.ts
