+++ test/commands/run.ts
