+++ test/commands/smoke-test-list.ts
