+++ test/commands/smoke-test.ts
