+++ webpack.config.test.js
@@ -1,21 +1,32 @@
 const { readdirSync, statSync } = require("fs");
-const webpack = require("webpack");
 const { join: joinPath, sep, extname } = require("path");
 
 function getIntegrationTestFiles() {
-  const dirPath = joinPath(__dirname, "test-browser", "integration");
-  const files = readdirSync(dirPath);
-  return files
+  let hlcDirPath = joinPath(__dirname, "test-browser", "integration");
+  let hlcFiles = readdirSync(hlcDirPath);
+  let rlcDirPath = joinPath(__dirname, "test-browser", "rlcIntegration");
+  let rlcFiles = readdirSync(rlcDirPath);
+  hlcFiles = hlcFiles
     .filter(
       name =>
-        extname(name) === ".js" && statSync(`${dirPath}${sep}${name}`).isFile()
+        extname(name) === ".js" && statSync(`${hlcDirPath}${sep}${name}`).isFile()
     )
-    .map(filename => `${dirPath}${sep}${filename}`);
+    .map(filename => `${hlcDirPath}${sep}${filename}`);
+
+  rlcFiles = rlcFiles
+    .filter(
+      name =>
+        extname(name) === ".js" && statSync(`${rlcDirPath}${sep}${name}`).isFile()
+    )
+    .map(filename => `${rlcDirPath}${sep}${filename}`);
+
+  return [...hlcFiles, ...rlcFiles];
 }
 
 const entry = getIntegrationTestFiles();
 
 module.exports = {
+  target: "web",
   entry,
   output: {
     filename: "index.js",
@@ -31,20 +42,5 @@ module.exports = {
     ]
   },
   mode: "development",
-  devtool: "inline-source-map",
-  resolve: {
-    fallback: {
-      fs: false,
-      path: require.resolve("path-browserify"),
-      buffer: require.resolve("buffer")
-    }
-  },
-  plugins: [
-    new webpack.ProvidePlugin({
-      Buffer: ["buffer", "Buffer"]
-    }),
-    new webpack.ProvidePlugin({
-      process: "process/browser"
-    })
-  ]
+  devtool: "inline-source-map"
 };
